fix txm tests
